% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saianitha02.R
\name{cir_of_circle}
\alias{cir_of_circle}
\title{Provides circumference of a circle}
\usage{
cir_of_circle(radius)
}
\arguments{
\item{radius}{radius of a circle}
}
\value{
cof circumference of a circle
}
\description{
Provides circumference of a circle
}
